+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-Nothing yet...
+### Changed
+- Allow several `TypeConfiguration` impls for the same model. [#18](https://github.com/mrahhal/MR.Augmenter/issues/18)
+- A more robust collecting of TypeConfiguration implementations.
 
 ## [1.0.0] - 2017-12-14
 ### Changed