+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+Nothing yet...
+
+## [1.0.2] - 2018-01-05
 ### Changed
 - Allow several `TypeConfiguration` impls for the same model. [#18](https://github.com/mrahhal/MR.Augmenter/issues/18)
 - A more robust collecting of TypeConfiguration implementations.
@@ -15,5 +18,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Support Asp.Net Core 2.0.
 
-[Unreleased]: https://github.com/mrahhal/MR.AspNetCore.Jobs/compare/1.0.0...HEAD
-[1.0.0]: https://github.com/mrahhal/MR.AspNetCore.Jobs/compare/1.0.0...0.3.1
+[Unreleased]: https://github.com/mrahhal/MR.AspNetCore.Jobs/compare/1.0.2...HEAD
+[1.0.2]: https://github.com/mrahhal/MR.AspNetCore.Jobs/compare/1.0.0...1.0.2
+[1.0.0]: https://github.com/mrahhal/MR.AspNetCore.Jobs/compare/0.3.1...1.0.0